API working and index.php working

+++ api/index.php
@@ -43,7 +43,7 @@
 * 
 * TICKETS
 *   /ticket/{id} return the ticket information IF the user has access to the project
-                 equivalent to /project/{id}/ticket/{id_simple_ticket}. all the following can be used on both path
+*                equivalent to /project/{id}/ticket/{id_simple_ticket}. all the following can be used on both path
 *   /ticket/{id}/comments get the comments of the ticket
 *   /ticket/{id}/comment/{id_comment} return the comment detail
 *   /ticket/{id}/comment/{id_comment}/remove
@@ -60,8 +60,6 @@
 *    -comment
 * 
 **/
-
-
 require_once "router.php";
 require_once "../functions.php";
 
@@ -95,11 +93,12 @@ function post($str, $optionnal = false){
 }
 
 /**
-* generate an error and clsoe the programm
+* generate an error and close the programm
 * @param code the http code
 * @param msg the message you want to display
 * @param args additionnal output you may want to display
 *
+* call a exit at the end of the function.
 */
 function http_error($code, $msg, $args = array()){
     http_response_code($code);
@@ -118,19 +117,42 @@ function http_error($code, $msg, $args = array()){
 }
 
 /**
+* generate a json report and close the programm
+* @param content an array
+* @param args additionnal output you may want to display
+*
+* call a exit at the end of the function.
+*/
+function http_success($content, $args = array()){
+    http_response_code(200);
+    $output = array(
+        "status"=>200,
+        "result"=>"ok",
+        "content"=>$content
+    );
+
+    foreach ($args as $t=>$v){
+        $output["$t"] = $v;
+    }
+
+    echo json_encode($output);
+    exit;
+}
+
+
+/**
 * return the user id if he is connected
 * kill the process otherwise
 **/
 function force_auth(){
     if (isset($_SESSION["user_id"])){
-        return $_SESSION["user_id"];
+        return (int) $_SESSION["user_id"];
     }else{
         http_error(401, "You need to be authentified to access this method");
     }
 }
 
 $route = new Route();
-function route(...$args){global $route;$route->route(...$args);};
 
 /*-----------------------------------------------------------------------------------------------------------------------------------------*/
 
@@ -138,7 +160,7 @@ function route(...$args){global $route;$route->route(...$args);};
 * /login login a user - return truthy or falsey depending on if the user successfully connected
 *     - POST:email
 *     - POST:password
-* /disconnext disconnect a user - clear all cookies and delete the session
+* /disconnect disconnect a user - clear all cookies and delete the session
 *
 **/
 $route->post(array("/api/login",
@@ -148,9 +170,11 @@ $route->post(array("/api/login",
 
     if (validate_user($mail, $password)){
         $_SESSION["user_id"] = get_user_by_email($mail)["id"];
-        echo "user logged in";
-    }else{    
-        echo "login failed";
+        $output = array("user_id"=>$_SESSION["user_id"]);
+
+        http_success($output);
+    }else{        
+        http_error(401, "login failed");
     }
 });
 
@@ -187,21 +211,20 @@ $route->post(array("/api/user/new",
     $id_user = add_user($name, $email, $password);
 
     $output = array("id_user"=>$id_user);
-    echo json_encode($output);
+    http_success($output);
 });
 
-//todo : check this function
-route("/api/user/me", function(){    
+$route->route("/api/user/me", function(){    
     $id = force_auth();
 
     $output = get_user($id);
-    echo json_encode($output);
+    http_success($output);
 });
 
 $route->get("/api/user/{id}", function($id){
     $id = (int) $id;
     $output = get_user($id);
-    echo json_encode($output);
+    http_success($output);
 });
 
 /** edit the user info - POST
@@ -209,14 +232,19 @@ $route->get("/api/user/{id}", function($id){
 * optional parameter : password, name, email
 *
 * @return the user infos
+*
+* you can only edit your own profile
 **/
-//todo : add a security
 $route->post(array("/api/user/{id}/edit",
                    "/api/user/me/edit"), function($id = null){
     $id = ($id !== null) ? (int) $id : force_auth();
     $name = post("name", true);
     $email = post("email", true);
     $password = post("password", true);
+
+    if ($id != force_auth()){
+        http_error(403, "You cannot edit another user's profile");
+    }
 
     $args = array(":id"=>$id);
     $set = array();
@@ -243,37 +271,38 @@ $route->post(array("/api/user/{id}/edit",
 
 
     $output = get_user($id);
-    echo json_encode($output);
+    http_success($output);
 });
 
+/**
+* Delete the user account
+* only the logged in user can delete his own account.
+*
+**/
 $route->delete(array("/api/user/me/delete",
                      "/api/user/{id}/delete"),
-               function($id = 0){
-                   $id = $id !== 0 ? (int) $id : force_auth();
+               function($id = null){
+                   $id = ($id !== null) ? (int) $id : force_auth();
 
-                   //todo : add more security
-                   if (isset($_SESSION["user_id"]) && $id == $_SESSION["user_id"]){
+                   if ($id == force_auth()){
                        delete_user($id);
                        if ($id == $_SESSION["user_id"]){
                            session_destroy();
                        }
                    }else{
-                       echo $id." - ".$_SESSION["user_id"]."\n";
-                       echo "you cannot destroy an account if it is not yours / you are not connected";
+                       http_error(403, "you cannot destroy an account if it is not yours / you are not connected.");
                    }
-
                    $output = array();
-                   echo json_encode($output);
+                   http_success($output);
                });
 
-//todo check here
-$route->route(array("/api/user/{id}/projects",
-                    "/api/user/me/projects",
+$route->route(array("/api/user/me/projects",
+                    "/api/user/{id}/projects",
                     "/api/projects"), function($id = null){
 
     $id = ($id === null) ? force_auth() : (int) $id;
     $output = get_projects_for_user($id);
-    echo json_encode($output);
+    http_success($output);
 });
 
 /**
@@ -314,14 +343,14 @@ $route->post("/api/project/new", function(){
     $id_project = add_project($name,$id_user, $ticket_prefix);
 
     $output = array("id_project"=>$id_project);
-    echo json_encode($output);
+    http_success($output);
 });
 
 
 $route->get("/api/project/{id}", function($id){
     $id = (int) $id;
     $project = get_project($id);
-    echo json_encode($project);
+    http_success($project);
 });
 
 /*only the creator can change these parameters for now*/
@@ -350,7 +379,7 @@ $route->post("/api/project/{id}/edit", function($id){
     }
 
     $output = get_project($id);
-    echo json_encode($output);
+    http_success($output);
 });
 
 /*only the creator can delete for now*/
@@ -360,62 +389,115 @@ $route->delete("/api/project/{id}/delete", function($id){
 
     if (is_admin($id_user,$id)){
         delete_project($id);
+        $output = array("delete"=>"ok");
+        http_success($output);
     }else{
-        echo "error - you do not have the right access level to do that";
+        http_error(403, "You need to be the admin of this project to be able to delete it", array("delete"=>"fail"));
     }
 });
 
-//todo : add user verification and add a return
 $route->post("/api/project/{id}/adduser", function($id_project){
-    $id_project = (int) $id_project;
-    //todo : add verification that the logged user CAN do it
+    $id_project = (int) $id_project; 
+
     $id_user = (int) post("id_user");
     $access_level = (int) post("access_level");
 
-    add_link_user_project($id_user, $id_project, $access_level);    
+    $id_current_user = force_auth();
+    if (! is_admin($id_current_user, $id_project)){
+        http_error(403, "Only an admin can add users to a project.");
+    }
+
+    add_link_user_project($id_user, $id_project, $access_level); 
+
+    http_success(array("link_user_project"=>get_link_user_project($id_user, $id_project)));
 });
 
 /*return the users on the project*/
-//todo : verify if the user has the right to see that
-$route->get("/api/project/{id}/users", function($id){
-    $id = (int) $id;
-    echo json_encode(get_users_for_project($id));
+$route->get("/api/project/{id}/users", function($id_project){
+    $id_project = (int) $id_project;
+    $id_user = force_auth();
+
+    if (access_level($id_user, $id_project) > 0){
+        http_success(get_users_for_project($id));
+    }else{
+        http_error(403, "You cannot see the users of a project you are not a part of.");
+    }
 });
 
-//todo : verify the user CAN do it
+/**
+* remove an user from a project
+* note that only an admin can remove other users, and only the creator can remove admins.
+**/
 $route->post("/api/project/{id}/removeuser", function($id_project){
     $id_project = (int) $id_project;
-    //todo : add verification that the logged user CAN do it
     $id_user = (int) post("id_user");
+    $actual_user = force_auth();
 
-    delete_link_user_project($id_user, $id_project);
+    if ($actual_user == $id_user || (is_admin($actual_user, $id_project) && access_level($actual_user, $id_project) > access_level($id_user, $id_project))){
+        delete_link_user_project($id_user, $id_project);
+        http_success(array("delete"=>"ok"));
+    }else{
+        http_error(403, "only an admin or the actual user can do that.");
+    }
 });
 
-//todo : verify theuser can do it
 $route->post("/api/project/{id}/addticket", function($id_project){
     $id_project = (int) $id_project;
-    $title = post("title");
+    $title = post("name");
     $priority = post("priority");
     $description = post("description");
     $due_date = post("due_date");
+    $manager_id = post("manager_id");
 
     $creator_id = force_auth();
 
-    $id = add_ticket($title, $id_project, $creator_id, $creator_id, $priority, $description, $due_date);
+    /*verify the user has the right to create tickets*/
+    if (access_level($creator_id, $id_project) >= 3){
+        $id = add_ticket($title, $id_project, $creator_id, $manager_id, $priority, $description, $due_date);
+        $output = array("id_ticket" => $id);
+        http_success($output); 
+    }else{
+        http_error(403, "You do not have the permission to create a ticket on this project");
+    }
 
-    $output = array("id_ticket" => $id);
-    echo json_encode($output);    
 });
 
-//todo : add false case and vverify if user has the right to check this
 $route->get("/api/project/{id}/tickets", function($id_project){
-    echo json_encode(get_tickets_for_project($id_project));
+    $id_project = (int) $id_project;
+    $current_user = force_auth();
+
+    if (get_project(get_project)){ 
+        if (access_level($current_user, $id_project) >= 1){
+            http_success(get_tickets_for_project($id_project));
+        }else{
+            http_error(403, "You do not have the right to access this project or his tickets");
+        }     
+    }else{
+        http_error(412, "This project does not exist");
+    }
 });
 
-//todo : add false case
 $route->get("/api/project/{id_project}/ticket/{id_simple_ticket}", function($id_project, $id_simple_ticket){
-    echo json_encode(get_ticket_simple($id_project, $id_simple_ticket));
+    $id_project = (int) $id_project;
+    $current_user = force_auth();
+
+    if (get_project(get_project)){ 
+        if (access_level($current_user, $id_project) >= 1){
+            http_success(get_tickets_for_project($id_project));
+            $ticket = get_ticket_simple($id_project, $id_simple_ticket);
+            if ($ticket){
+                http_success($ticket);
+            }else{
+                http_error(412, "This ticket does not exists within the project.");
+            }
+        }else{
+            http_error(403, "You do not have the right to access this project or his tickets");
+        }     
+    }else{
+        http_error(412, "This project does not exist");
+    }
 });
+
 
 /**
 * TICKETS
@@ -427,57 +509,80 @@ $route->get("/api/project/{id_project}/ticket/{id_simple_ticket}", function($id_
 *      -comment
 *      user needs to be authenticated
 **/
-
 $route->get("/api/ticket/{id}", function($id){
-    echo json_encode(get_ticket($id));
+    $access_level = rights_user_ticket(force_auth(), $id);
+    if ($access_level >=1){
+        http_success(get_ticket($id));
+    }else{
+        http_error(403, "You do not have the right to se this ticket");
+    }
 });
 
-/*todo : check the user can do that*/
 $route->post("/api/ticket/{id}/edit", function($id_ticket){
     $id_ticket = (int) $id_ticket;
+    $ticket = get_ticket($id_ticket);
+
     $params = array(
         "name"=>post("name", true),
         "priority"=>post("priority", true),
         "description"=>post("description", true),
-        "due_date"=>post("due_date", true)
+        "due_date"=>post("due_date", true),
+        "manager_id"=>post("manager_id", true)
     );
 
-    $args = array(":ticket_id"=>$id_ticket);
-    $set = array();
+    $access_level = rights_user_ticket(force_auth(), $id_ticket);
 
-    foreach($params as $t=>$v){
-        if ($v){
-            $args[":$t"] = $v;
-            $set[] = "$t = :$t";
+    /*verify the user has the right to edit the ticket*/
+    /*creator can modify / manager can modify without modifying himself*/
+    if ($access_level >= 4 || ($access_level == 3 && !isset($params["manager_id"]))){
+
+        $args = array(":ticket_id"=>$id_ticket);
+        $set = array();
+
+        foreach($params as $t=>$v){
+            if ($v){
+                $args[":$t"] = $v;
+                $set[] = "$t = :$t";
+            }
         }
+
+        if (count($set) >= 1){
+            $req = "UPDATE tickets SET ".join(",",$set)." WHERE id=:ticket_id;";
+            execute($req, $args);
+        }
+
+        http_success(get_ticket($id_ticket));
+    }else{
+        http_error(403, "You do not have the permission to modify this ressource");
     }
 
-    if (count($set) >= 1){
-        $req = "UPDATE tickets SET ".join(",",$set)." WHERE id=:ticket_id;";
-        execute($req, $args);
-    }
-
-    echo json_encode(get_ticket($id_ticket));
+    ;
 });
 
-/*todo : check the user can do that*/
 $route->delete("/api/ticket/{id}/delete", function($id_ticket){
-    delete_ticket($id_ticket);
+    if (rights_user_ticket(force_auth(), $id_ticket) >= 4){
+        delete_ticket($id_ticket);
+        http_success(array("delete"=>"ok"));
+    }else{
+        http_error(403,"You do not have the right to do that");
+    }
 });
 
-/*todo : check the user can do that AND that the ticket exists*/
 $route->post("/api/ticket/{id}/addcomment", function($ticket_id){
     $ticket_id = (int) $ticket_id;
     $comment = post("comment");
     $creator_id = force_auth();
-
-    $id = add_comment($ticket_id, $creator_id, $comment);
-    $output = array("id_comment"=>$id);
-    echo json_encode($output);
+    if (rights_user_ticket($creator_id, $ticket_id) >= 2){
+        $id = add_comment($ticket_id, $creator_id, $comment);
+        $output = array("id_comment"=>$id);
+        http_success($output);
+    }else{
+        http_error(403, "You do not have the permission to comment");
+    }
 });
 
 $route->get("/api/ticket/{id}/comments", function($id){
-    echo json_encode(get_comments_for_ticket($id));
+    http_success(get_comments_for_ticket($id));
 });
 
 
@@ -489,22 +594,24 @@ $route->get("/api/ticket/{id}/comments", function($id){
 * DELETE /api/comment/{id}/delete
 *
 **/
-//todo : check if the user has rights
 $route->get("/api/comment/{id_comment}", function($id_comment){
-    echo json_encode(get_comment($id_comment));
+    if (rights_user_comment(force_auth(), $id_comment) > 0){
+        http_success(get_comment($id_comment));
+    }else{
+        http_error(403, "You do not have the permission to access this comment");
+    }
 });
 
 $route->post("/api/comment/{id}/edit", function($comment_id){
     $comment = post("comment");
-    $creator_id = force_auth();
+    $user = force_auth();
 
-    if (get_comment($comment_id)["creator_id"] != $creator_id){
-        http_error(403,"You cannot modify this comment as it isn't yours");
+    if (rights_user_comment($user, $comment_id) !== 2){
+        http_error(403,"You cannot modify this comment");
+    }else{
+        $id = edit_comment($comment_id, $comment);
+        http_success(get_comment($comment_id));
     }
-
-    $id = edit_comment($comment_id, $comment);
-
-    echo json_encode(get_comment($comment_id));
 });
 
 $route->delete("/api/comment/{id}/delete", function($comment_id){
@@ -512,9 +619,12 @@ $route->delete("/api/comment/{id}/delete", function($comment_id){
     $id_user = force_auth();
     $id_project = execute("SELECT project_id FROM tickets WHERE id IN (SELECT ticket_id FROM comments WHERE id = ?)", array($comment_id))->fetch()["project_id"];
 
-    if (is_admin($id_user, $id_project) || get_comment($comment_id)["creator_id"] == $creator_id){
+    if (rights_user_comment($id_user, $comment_id) === 2){
         delete_comment($comment_id);
-    }
+        http_success(array("delete"=>"ok"));
+    }else{
+        http_error(403, "You cannot delete this comment as it is not yours / you are not admin");
+    }    
 });
 
 /**
@@ -523,12 +633,6 @@ $route->delete("/api/comment/{id}/delete", function($comment_id){
 *
 **/
 $route->error_404(function(){
-    http_response_code(404);
-    $output = array(
-        "error_code" => 404,
-        "message" => "404 - the server couldn't find a page corresponding to your url and method settings"
-    );
-    echo json_encode($output);
-    exit;
+    http_error(404, "Error 404 - The server cannot find a page corresponding to your request. please check your url and method.");
 });
 ?>
